[refactor] components/[Feature별 디렉토리]/shared/ui -> ui 레이어 제거
