+++ src/app/auth/signup/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from 'next';
-import AuthPoster from '@/components/auth/shared/ui/AuthPoster';
+import AuthPoster from '@/components/auth/shared/AuthPoster';
 import SignUpForm from '@/components/auth/SignUpForm';
 
 export const metadata: Metadata = {
