+++ src/components/auth/ProfileForm.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ProfileEditDialog from '@/components/shared/ui/ProfileEditDialog';
+import ProfileEditDialog from '@/components/shared/ProfileEditDialog';
 
 export default function ProfileForm() {
     return <ProfileEditDialog setIsProfileEditDialogOpen={() => { }} />
