+++ src/components/gatherings/detail/Information.tsx
@@ -4,9 +4,9 @@ import { formatDate, formatTime } from '@/utils/shared/date';
 import { Gathering, Participant } from '@/types/gatherings';
 import { Check, UserRoundCheck } from "lucide-react"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/shadcn-ui/tooltip";
-import ImageWithFallback from '@/components/shared/ui/ImageWithFallback';
-import SaveToggleButton from '@/components/gatherings/shared/ui/SaveToggleButton';
-import JoinedCountsProgressBar from '@/components/gatherings/shared/ui/JoinedCountsProgressBar';
+import ImageWithFallback from '@/components/shared/ImageWithFallback';
+import SaveToggleButton from '@/components/gatherings/shared/SaveToggleButton';
+import JoinedCountsProgressBar from '@/components/gatherings/shared/JoinedCountsProgressBar';
 
 /** 모임 상세 페이지 상단 우측 정보 */
 export default function Information({ detail, id, participants }: { detail: Gathering, id: string, participants: Participant[] }) {
