+++ src/components/gatherings/shared/JoinedCountsProgressBar.tsx
