+++ src/components/mypage/CreatableReview.tsx
@@ -1,8 +1,8 @@
 import { useGatheringsStore } from '@/store/gatheringsStore';
 import { Gathering } from '@/types/gatherings';
-import ImageWithFallback from '../shared/ui/ImageWithFallback';
-import GatheringInformation from '@/components/mypage/shared/ui/GatheringInformation';
-import Button from '@/components/shared/ui/Button';
+import ImageWithFallback from '@/components/shared/ImageWithFallback';
+import GatheringInformation from '@/components/mypage/shared/GatheringInformation';
+import Button from '@/components/shared/Button';
 
 /** 나의 리뷰 - 작성 가능한 리뷰 */
 export default function CreatableReview({ gathering, myReviewsTab, userId, onOpenReviewDialog }: { gathering: Gathering, myReviewsTab: number, userId: number, onOpenReviewDialog: (gathering: { userId: number, gatheringId: number }) => void }) {
