+++ src/components/mypage/CreatedGatherings.tsx
@@ -5,11 +5,11 @@ import { useContext } from 'react';
 import { AuthContext } from '@/providers/AuthProvider';
 import { getTimeRemaining } from '@/utils/shared/date';
 import dynamic from 'next/dynamic';
-import ImageWithFallback from '@/components/shared/ui/ImageWithFallback';
-import GatheringInformation from '@/components/mypage/shared/ui/GatheringInformation';
-import DateReminder from '@/components/shared/ui/DateReminder'
+import ImageWithFallback from '@/components/shared/ImageWithFallback';
+import GatheringInformation from '@/components/mypage/shared/GatheringInformation';
+import DateReminder from '@/components/shared/DateReminder'
 
-const LoadingUI = dynamic(() => import('@/components/mypage/shared/ui/LoadingUI'), { ssr: false });
+const LoadingUI = dynamic(() => import('@/components/mypage/shared/LoadingUI'), { ssr: false });
 
 /** 마이페이지 '내가 만든 모임' */
 export default function CreatedGatherings() {
