+++ src/components/mypage/JoinedGatherings.tsx
@@ -7,14 +7,14 @@ import { AuthContext } from '@/providers/AuthProvider';
 import { getTimeRemaining, toKoreanTime } from '@/utils/shared/date';
 import { CheckCircle } from "lucide-react"
 import dynamic from 'next/dynamic';
-import Button from '@/components/shared/ui/Button';
-import GatheringInformation from '@/components/mypage/shared/ui/GatheringInformation';
-import DateReminder from '@/components/shared/ui/DateReminder';
-import ImageWithFallback from '@/components/shared/ui/ImageWithFallback';
+import Button from '@/components/shared/Button';
+import GatheringInformation from '@/components/mypage/shared/GatheringInformation';
+import DateReminder from '@/components/shared/DateReminder';
+import ImageWithFallback from '@/components/shared/ImageWithFallback';
 
-const LoadingUI = dynamic(() => import('@/components/mypage/shared/ui/LoadingUI'), { ssr: false });
-const ConfirmDialog = dynamic(() => import('@/components/shared/ui/ConfirmDialog'), { ssr: false });
-const OverlayForDisabled = dynamic(() => import('@/components/shared/ui/OverlayForDisabled'), { ssr: false });
+const LoadingUI = dynamic(() => import('@/components/mypage/shared/LoadingUI'), { ssr: false });
+const ConfirmDialog = dynamic(() => import('@/components/shared/ConfirmDialog'), { ssr: false });
+const OverlayForDisabled = dynamic(() => import('@/components/shared/OverlayForDisabled'), { ssr: false });
 
 interface JoinedGatheringsProps {
   setSelectedTab: (tab: number) => void;
