+++ src/components/mypage/shared/GatheringInformation.tsx
