+++ src/components/shared/ConfirmDialog.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { X } from 'lucide-react';
-import Button from './Button';
+import Button from '@/components/shared/Button';
 
 interface CheckingModalProps {
     isOpen: boolean;
