+++ src/utils/gatherings/gatheringsParamUtils.ts
@@ -1,5 +1,5 @@
 import { getApiTypeFromMainType } from '@/utils/gatherings/gatheringsTypeUtils';
-import { GATHERING_CONSTANTS } from '@/components/gatherings/shared/constants/gatheringsConstants';
+import { GATHERING_CONSTANTS } from '@/utils/gatherings/constants/gatheringsConstants';
 
 /**
  * 모임 API 파라미터 옵션
